This adds a new bpf map cilium_lb4_backends_v2
The V2 map key is u32 typed instead of u16.

Upon agent (re)start:
- we restore backends from v1 map and copy all
entries from v1 map to v2 map if v1 map exists
(this is the upgrade scenario).
- we then remove v1 map and operate on v2 map.

+++ pkg/maps/lbmap/maglev_outer_map.go
@@ -104,7 +104,7 @@ func MaglevOuterMapTableSize(mapName string) (bool, uint32) {
 	}
 	defer innerMap.Close()
 
-	return true, innerMap.ValueSize() / uint32(unsafe.Sizeof(uint16(0)))
+	return true, innerMap.ValueSize() / uint32(unsafe.Sizeof(uint32(0)))
 }
 
 // Update updates the value associated with a given key for a maglev outer map.
